if case for teardown

+++ acceptance-tests/src/test/java/tech/pegasys/web3signer/tests/publickeys/AwsKeyIdentifiersAcceptanceTest.java
@@ -63,8 +63,10 @@ public class AwsKeyIdentifiersAcceptanceTest extends KeyIdentifiersAcceptanceTes
 
   @AfterAll
   void teardown() {
-    awsSecretsManagerUtil.deleteSecret();
-    awsSecretsManagerUtil.close();
+    if (awsSecretsManagerUtil != null) {
+      awsSecretsManagerUtil.deleteSecret();
+      awsSecretsManagerUtil.close();
+    }
   }
 
   @Test
